manejo de errores en UsuarioDAOH2.kt

+++ src/main/kotlin/data/dao/UsuarioDAOH2.kt
@@ -15,22 +15,39 @@ class UsuarioDAOH2 : IUsuarioDAO {
      * ejecuta la consulta y seguidamente maneja cualquier error que pueda surgir.
      */
     override fun insertarCampo(nombre: String, email: String) {
-        val connection = DatabaseTienda.getConnection()
-        var stmt: Statement? = null
-
-        try{
-            val sql = "INSERT INTO Usuario (nombre, email) VALUES (?, ?)"
-            stmt = connection.prepareStatement(sql)
-            stmt.setString(1, nombre)
-            stmt.setString(2, email)
-            stmt.executeUpdate()
+        val connection = try {
+            DatabaseTienda.getConnection()
         } catch (e: SQLException) {
-            throw SQLException("Error al insertar los campos en las tablas", e)
+            throw SQLException("Error al obtener la conexión: ${e.message}")
+            null
         } catch (e: Exception) {
-            throw Exception("Error: ${e.message}")
-        } finally {
-            stmt?.close()
-            DatabaseTienda.closeConnection(connection)
+            throw Exception("Error inesperado: ${e.message}")
+            null
+        }
+        println("Conexión exitosa.")
+        if (connection != null){
+            var stmt: Statement? = null
+            try {
+                val sql = "INSERT INTO Usuario (nombre, email) VALUES (?, ?)"
+                stmt = connection.prepareStatement(sql)
+                stmt.setString(1, nombre)
+                stmt.setString(2, email)
+                stmt.executeUpdate()
+            } catch (e: SQLException) {
+                throw SQLException("Error al insertar los campos en las tablas: ${e.message}")
+            } catch (e: Exception) {
+                throw Exception("Error inesperado: ${e.message}")
+            } finally {
+                stmt?.close()
+                try{
+                    DatabaseTienda.closeConnection(connection)
+                    println("Se cerró la conexión exitosamente.")
+                } catch (e: SQLException) {
+                    throw SQLException("Error al cerrar la conexión: ${e.message}")
+                } catch (e: Exception) {
+                    throw Exception("Error inesperado: ${e.message}")
+                }
+            }
         }
     }
 
@@ -42,22 +59,39 @@ class UsuarioDAOH2 : IUsuarioDAO {
      * Funciona igual que el metodo anterior, pero más limpio porque se tiene una clase con los datos agrupados.
      */
     override fun insertarCampo(usuario: Usuario) {
-        val connection = DatabaseTienda.getConnection()
-        var stmt: Statement? = null
-
-        try{
-            val sql = "INSERT INTO Usuario (nombre, email) VALUES (?, ?)"
-            stmt = connection.prepareStatement(sql)
-            stmt.setString(1, usuario.nombre)
-            stmt.setString(2, usuario.correo)
-            stmt.executeUpdate()
+        val connection = try {
+            DatabaseTienda.getConnection()
         } catch (e: SQLException) {
-            throw SQLException("Error al insertar los campos en las tablas", e)
+            throw SQLException("Error al obtener la conexión: ${e.message}")
+            null
         } catch (e: Exception) {
-            throw Exception("Error: ${e.message}")
-        } finally {
-            stmt?.close()
-            DatabaseTienda.closeConnection(connection)
+            throw Exception("Error inesperado: ${e.message}")
+            null
+        }
+        println("Conexión exitosa.")
+        if (connection != null){
+            var stmt: Statement? = null
+            try {
+                val sql = "INSERT INTO Usuario (nombre, email) VALUES (?, ?)"
+                stmt = connection.prepareStatement(sql)
+                stmt.setString(1, usuario.nombre)
+                stmt.setString(2, usuario.correo)
+                stmt.executeUpdate()
+            } catch (e: SQLException) {
+                throw SQLException("Error al insertar los campos en las tablas: ${e.message}")
+            } catch (e: Exception) {
+                throw Exception("Error inesperado: ${e.message}")
+            } finally {
+                stmt?.close()
+                try{
+                    DatabaseTienda.closeConnection(connection)
+                    println("Se cerró la conexión exitosamente.")
+                } catch (e: SQLException) {
+                    throw SQLException("Error al cerrar la conexión: ${e.message}")
+                } catch (e: Exception) {
+                    throw Exception("Error inesperado: ${e.message}")
+                }
+            }
         }
     }
 
@@ -69,28 +103,47 @@ class UsuarioDAOH2 : IUsuarioDAO {
      * @throws Exception Para cualquier otro tipo de error inesperado.
      */
     override fun getAll(): List<Usuario> {
-        val conn = DatabaseTienda.getConnection()
-        val listaUsuarios = mutableListOf<Usuario>()
-        var stmt: PreparedStatement? = null
-        var rs: ResultSet? = null
-        try {
-            val sql = "SELECT * FROM Usuario"
-            stmt = conn.prepareStatement(sql)
-            rs = stmt.executeQuery()
-            while (rs.next()) {
-                val id = rs.getInt("id")
-                val nombre = rs.getString("nombre")
-                val email = rs.getString("email")
-                listaUsuarios.add(Usuario(id, nombre, email))
-            }
+        val conn = try {
+            DatabaseTienda.getConnection()
         } catch (e: SQLException) {
-            throw SQLException("Error al obtener los usuarios: ${e.message}")
-        } catch(e: Exception) {
-            throw Exception("Error: ${e.message}")
-        }finally {
-            rs?.close()
-            stmt?.close()
-            DatabaseTienda.closeConnection(conn)
+            throw SQLException("Error al obtener la conexión: ${e.message}")
+            null
+        } catch (e: Exception) {
+            throw Exception("Error inesperado: ${e.message}")
+            null
+        }
+        val listaUsuarios = mutableListOf<Usuario>()
+        println("Conexión exitosa.")
+        if (conn != null){
+            var stmt: PreparedStatement? = null
+            var rs: ResultSet? = null
+            try {
+                val sql = "SELECT * FROM Usuario"
+                stmt = conn.prepareStatement(sql)
+                rs = stmt.executeQuery()
+                while (rs.next()) {
+                    val id = rs.getInt("id")
+                    val nombre = rs.getString("nombre")
+                    val email = rs.getString("email")
+                    listaUsuarios.add(Usuario(id, nombre, email))
+                }
+            } catch (e: SQLException) {
+                throw SQLException("Error al obtener los usuarios: ${e.message}")
+            } catch(e: Exception) {
+                throw Exception("Error: ${e.message}")
+            }finally {
+                rs?.close()
+                stmt?.close()
+                try{
+                    DatabaseTienda.closeConnection(conn)
+                    println("Se cerró la conexión exitosamente.")
+                } catch (e: SQLException) {
+                    throw SQLException("Error al cerrar la conexión: ${e.message}")
+                } catch (e: Exception) {
+                    throw Exception("Error inesperado: ${e.message}")
+                }
+            }
+            return listaUsuarios
         }
         return listaUsuarios
     }
@@ -104,26 +157,45 @@ class UsuarioDAOH2 : IUsuarioDAO {
      * @throws Exception Para cualquier otro tipo de error inesperado.
      */
     override fun getUsuariosByProductoComprado(nombreProducto: String): List<Usuario> {
-        val conn = DatabaseTienda.getConnection()
-        val usuarios = mutableListOf<Usuario>()
-        var stmt: PreparedStatement? = null
-        var rs: ResultSet? = null
-        try {
-            val sql = "SELECT DISTINCT U.id, U.nombre, U.email FROM Usuario U JOIN Pedido P ON U.id = P.idUsuario JOIN LineaPedido LP ON P.id = LP.idPedido JOIN Producto PR ON LP.idProducto = PR.id WHERE PR.nombre = ?"
-            stmt = conn.prepareStatement(sql)
-            stmt.setString(1, nombreProducto)
-            rs = stmt.executeQuery()
-            while (rs.next()) {
-                usuarios.add(Usuario(rs.getInt("id"), rs.getString("nombre"), rs.getString("email")))
-            }
+        val conn = try {
+            DatabaseTienda.getConnection()
         } catch (e: SQLException) {
-            throw SQLException("Error al obtener usuarios que compraron '$nombreProducto': ${e.message}")
+            throw SQLException("Error al obtener la conexión: ${e.message}")
+            null
         } catch (e: Exception) {
-            throw Exception("Error: ${e.message}")
-        } finally {
-            rs?.close()
-            stmt?.close()
-            DatabaseTienda.closeConnection(conn)
+            throw Exception("Error inesperado: ${e.message}")
+            null
+        }
+        val usuarios = mutableListOf<Usuario>()
+        println("Conexión exitosa.")
+        if (conn != null){
+            var stmt: PreparedStatement? = null
+            var rs: ResultSet? = null
+            try {
+                val sql = "SELECT DISTINCT U.id, U.nombre, U.email FROM Usuario U JOIN Pedido P ON U.id = P.idUsuario JOIN LineaPedido LP ON P.id = LP.idPedido JOIN Producto PR ON LP.idProducto = PR.id WHERE PR.nombre = ?"
+                stmt = conn.prepareStatement(sql)
+                stmt.setString(1, nombreProducto)
+                rs = stmt.executeQuery()
+                while (rs.next()) {
+                    usuarios.add(Usuario(rs.getInt("id"), rs.getString("nombre"), rs.getString("email")))
+                }
+            } catch (e: SQLException) {
+                throw SQLException("Error al obtener usuarios que compraron '$nombreProducto': ${e.message}")
+            } catch (e: Exception) {
+                throw Exception("Error: ${e.message}")
+            } finally {
+                rs?.close()
+                stmt?.close()
+                try{
+                    DatabaseTienda.closeConnection(conn)
+                    println("Se cerró la conexión exitosamente.")
+                } catch (e: SQLException) {
+                    throw SQLException("Error al cerrar la conexión: ${e.message}")
+                } catch (e: Exception) {
+                    throw Exception("Error inesperado: ${e.message}")
+                }
+            }
+            return usuarios
         }
         return usuarios
     }
@@ -138,20 +210,39 @@ class UsuarioDAOH2 : IUsuarioDAO {
      * @throws Exception Si ocurre cualquier otro tipo de error inesperado.
      */
     override fun deleteByName(nombre: String){
-        var conn = DatabaseTienda.getConnection()
-        var stmt: PreparedStatement? = null
-        try{
-            val sql = "DELETE FROM Usuario WHERE nombre = ?"
-            stmt = conn.prepareStatement(sql)
-            stmt.setString(1, nombre)
-            stmt.executeUpdate()
+        val conn = try {
+            DatabaseTienda.getConnection()
         } catch (e: SQLException) {
-            throw SQLException("Error al eliminar usuario '$nombre' en la base de datos: ${e.message}")
+            throw SQLException("Error al obtener la conexión: ${e.message}")
+            null
         } catch (e: Exception) {
-            throw Exception("Error: ${e.message}")
-        } finally {
-            stmt?.close()
-            DatabaseTienda.closeConnection(conn)
+            throw Exception("Error inesperado: ${e.message}")
+            null
+        }
+        val usuarios = mutableListOf<Usuario>()
+        println("Conexión exitosa.")
+        if (conn != null){
+            var stmt: PreparedStatement? = null
+            try{
+                val sql = "DELETE FROM Usuario WHERE nombre = ?"
+                stmt = conn.prepareStatement(sql)
+                stmt.setString(1, nombre)
+                stmt.executeUpdate()
+            } catch (e: SQLException) {
+                throw SQLException("Error al eliminar usuario '$nombre' en la base de datos: ${e.message}")
+            } catch (e: Exception) {
+                throw Exception("Error: ${e.message}")
+            } finally {
+                stmt?.close()
+                try{
+                    DatabaseTienda.closeConnection(conn)
+                    println("Se cerró la conexión exitosamente.")
+                } catch (e: SQLException) {
+                    throw SQLException("Error al cerrar la conexión: ${e.message}")
+                } catch (e: Exception) {
+                    throw Exception("Error inesperado: ${e.message}")
+                }
+            }
         }
     }
 }
